+++ simplq/src/components/pages/Home/StaticInfos.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { CreateQButton, JoinQButton } from '../../common/Button';
 import Header from '../../common/Header';
 import styles from '../../../styles/homePage.module.scss';
+import { CreatorSlider, JoinerSlider } from '../../common/slider';
 
 export const Banner = () => (
   <div>
@@ -41,39 +42,14 @@ export const BenefitsInfo = () => (
 export const HowToCreate = () => (
   <div className={styles['create-queue']}>
     <h2>Create a Queue at the click of a button</h2>
-    <div>
-      <img src="/images/queue_creation.svg" alt="Create Queue" />
-      <div>
-        <h3>Create Queue</h3>
-        <p>Enter queue name and create a queue totally free of cost</p>
-      </div>
-    </div>
-    <div>
-      <h3>Share Queue</h3>
-      <p>Share the queue to your users for them to join</p>
-    </div>
-    <div>
-      <h3>Manage Queue</h3>
-      <p>Manage your queue seamlessly and efficiently</p>
-    </div>
+    <CreatorSlider />
   </div>
 );
 
 export const HowToJoin = () => (
   <div className={styles['join-queue']}>
     <h2>Join a queue without being there</h2>
-    <div>
-      <h3>Join Queue</h3>
-      <p>Get queue link and join from anywhere</p>
-    </div>
-    <div>
-      <h3>Wait for your turn</h3>
-      <p>Get notified when your turn is up without physically being in the queue</p>
-    </div>
-    <div>
-      <h3>Manage Queue</h3>
-      <p>Manage your queue seamlessly and efficiently</p>
-    </div>
+    <JoinerSlider />
   </div>
 );
